148 Move helper functions to the testUtils package
