+++ testutils/src/main/scala/za/co/absa/enceladus/testutils/HelperFunctions.scala
@@ -13,12 +13,12 @@
  * limitations under the License.
  */
 
-package za.co.absa.enceladus.testutils.datasetComparison
+package za.co.absa.enceladus.testutils
 
 import org.apache.log4j.{LogManager, Logger}
-import org.apache.spark.sql.{Column, DataFrame}
 import org.apache.spark.sql.functions.{expr, max}
-import org.apache.spark.sql.types._
+import org.apache.spark.sql.types.{ArrayType, StructField, StructType}
+import org.apache.spark.sql.{Column, DataFrame}
 
 import scala.collection.mutable
